refactor: move ProcessRunner to RunnerBarCore/Services (#922)
- Delete Sources/RunnerBar/Services/ProcessRunner.swift
- Add Sources/RunnerBarCore/Services/ProcessRunner.swift with public API
- No import changes needed in callers (covered by @_exported in Exports.swift)
