+++ Sources/RunnerBar/Services/ProcessRunner.swift
@@ -1,106 +1,4 @@
 // ProcessRunner.swift
 // RunnerBar
-import Foundation
-
-// MARK: - ProcessRunner
-
-/// Shared primitive for launching subprocesses with streaming output,
-/// optional stdin, optional working directory, and a DispatchWorkItem timeout.
-///
-/// Both `runGHProcess` (GitHubCLITransport) and `runScriptWithOutput`
-/// (RunnerLifecycleService) are thin wrappers around this type.
-enum ProcessRunner {
-    /// A value type representing Result.
-    struct Result {
-        /// The data constant.
-        let data: Data?
-        /// The exitCode constant.
-        let exitCode: Int32
-        /// Convenience: decoded UTF-8 string of `data`, or empty string.
-        var output: String { data.flatMap { String(data: $0, encoding: .utf8) } ?? "" }
-    }
-
-    /// Launches an executable synchronously.
-    /// ⚠️ Must be called from a background thread — blocks until exit or timeout.
-    ///
-    /// - Parameters:
-    ///   - executableURL: Full path to the executable.
-    ///   - arguments: Command-line arguments.
-    ///   - stdin: Optional data written to the process's standard input.
-    ///   - workingDirectory: Optional working directory for the process.
-    ///   - mergeStderr: When true, stderr is merged into stdout (default false).
-    ///   - timeout: Seconds before the process is force-terminated.
-    /// - Returns: A `Result` with the collected stdout data and exit code.
-    ///   `exitCode` is `Int32.max` on process-launch failure.
-    static func run(
-        executableURL: URL,
-        arguments: [String],
-        stdin: Data? = nil,
-        workingDirectory: URL? = nil,
-        mergeStderr: Bool = false,
-        timeout: TimeInterval = 20
-    ) -> Result {
-        let task = Process()
-        task.executableURL = executableURL
-        task.arguments = arguments
-        if let workingDirectory { task.currentDirectoryURL = workingDirectory }
-
-        let outPipe = Pipe()
-        task.standardOutput = outPipe
-        task.standardError = mergeStderr ? outPipe : Pipe()
-
-        // Wire stdin pipe when body data is provided.
-        let inputPipe: Pipe?
-        if stdin != nil {
-            let p = Pipe()
-            task.standardInput = p
-            inputPipe = p
-        } else {
-            inputPipe = nil
-        }
-
-        nonisolated(unsafe) var outputData = Data()
-        let lock = NSLock()
-        outPipe.fileHandleForReading.readabilityHandler = { handle in
-            let chunk = handle.availableData
-            guard !chunk.isEmpty else { return }
-            lock.lock(); outputData.append(chunk); lock.unlock()
-        }
-
-        do {
-            try task.run()
-        } catch {
-            log("ProcessRunner › launch error: \(error) — \(executableURL.lastPathComponent) \(arguments.joined(separator: " "))")
-            outPipe.fileHandleForReading.readabilityHandler = nil
-            return Result(data: nil, exitCode: Int32.max)
-        }
-
-        // Write stdin after launch so the process is ready to consume it.
-        if let inputPipe, let stdinData = stdin {
-            inputPipe.fileHandleForWriting.write(stdinData)
-            inputPipe.fileHandleForWriting.closeFile()
-        }
-
-        let timeoutItem = DispatchWorkItem {
-            // Guard against the race where the process exits just before the
-            // timeout fires — only terminate if the process is still running.
-            guard task.isRunning else {
-                log("ProcessRunner › timeout fired but process already exited — \(executableURL.lastPathComponent)")
-                return
-            }
-            log("ProcessRunner › timeout (\(timeout)s) — terminating \(executableURL.lastPathComponent)")
-            task.terminate()
-        }
-        DispatchQueue.global().asyncAfter(deadline: .now() + timeout, execute: timeoutItem)
-        task.waitUntilExit()
-        timeoutItem.cancel()
-
-        outPipe.fileHandleForReading.readabilityHandler = nil
-        let tail = outPipe.fileHandleForReading.readDataToEndOfFile()
-        if !tail.isEmpty { lock.lock(); outputData.append(tail); lock.unlock() }
-
-        let exitCode = task.terminationStatus
-        log("ProcessRunner › exit=\(exitCode) bytes=\(outputData.count) — \(executableURL.lastPathComponent)")
-        return Result(data: outputData.isEmpty ? nil : outputData, exitCode: exitCode)
-    }
-}
+// Moved to RunnerBarCore/Services/ProcessRunner.swift — see #922.
+// This file is intentionally empty; the type is re-exported via Exports.swift.
